requirements.txt
Pillow=2.3.0
pyaml=15.8.2
click=6.6
